now with geo-finding!
the pins page now gets its pins from a /pins/within which needs a
searchArea which contains a geojson of type 'Polygon'.  the webserver is
now supplying a hard coded one.  Note that the 'Polygon' coordinates is
an array of arrays of [x,y] points where the last point is equal to the
first point (thus a polygon).  The first item in the outer array is the
actual polygon with the other items in the array being 'hole' polygons.
clear as mud?